% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the greatest common divisor of two numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{a, b}{numeric scalars or integers.}
}
\value{
A number that is the greatest common divisor of a and b.
}
\description{
This is an efficient function for computing the greatest common divisor (GCD) of two integers (numbers) a and b.
The arguments a and b can not be 0 because 0 has no divisor.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
[link url:](https://en.wikipedia.org/wiki/Euclidean_algorithm)
}
