% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\arguments{
\item{v1, v2}{vertices}

\item{w}{weight of the edge between two vertices}
}
\description{
(v1=1,v2=2,w=7) means that the length of the edge linking vertex 1 and vertex 2 is 7.
(v1=6,v2=5,w=9) means that the length of the edge linking vertex 6 and vertex 5 is 9.
}
\references{
/url:https://en.wikipedia.org/wiki/Graph
}
