% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Data to showcase graph for Dijkstras algorithm}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{v1}{a vertex}
\item{v2}{a vertex}
\item{w}{the weight of the edge linking v1 and v2}}
}
\source{
{Created in-house to serve as an example}
}
\usage{
wiki_graph
}
\description{
(v1=1,v2=2,w=7) means that the length of the edge linking vertex 1 and vertex 2 is 7.
(v1=6,v2=5,w=9) means that the length of the edge linking vertex 6 and vertex 5 is 9.
}
\examples{
data(wiki_graph)
}
\references{
https://en.wikipedia.org/wiki/Graph
}
\keyword{datasets}
